- refactor 2.0X processors, create shared base class - implement value mapping filtering on 2.0X - Fix AMP processor sector_percentage issue when percentage does not add up to 100

+++ import-core/import-services/src/main/java/org/devgateway/importtool/services/processor/AMPStaticProcessor.java
@@ -683,7 +683,7 @@ public class AMPStaticProcessor implements IDestinationProcessor {
 			JsonBean bean = new JsonBean();
 			bean.set(mapping.getSourceField().getFieldName() + "_id", entry.getKey());
 			if (mapping.getSourceField().isPercentage() && entry.getValue() > 0){
-				bean.set(mapping.getSourceField().getFieldName() + "_percentage", 100 / divider);
+				bean.set(mapping.getSourceField().getFieldName() + "_percentage", (double)100 / (double)divider);
 			}
 			beanList.add(bean);
 		}
